Move domain registrations card related classes to cards package
